requirements.txt
anthropic>=0.40.0
elevenlabs>=1.0.0
fpdf2>=2.7.0
Pillow>=10.0.0
python-dotenv>=1.0.0
